Move arb generation from bin to lib
